rename: 입력 모델을 별도의 패키지로 이동
