+++ src/main/kotlin/bingsoochef/bingsoochef/bingsoo/application/command/CreateBingsooCommand.kt
@@ -1,8 +1,8 @@
-package bingsoochef.bingsoochef.bingsoo.application
+package bingsoochef.bingsoochef.bingsoo.application.command
 
 import bingsoochef.bingsoochef.bingsoo.domain.Taste
 
-data class UpdateBingsooCommand(
+data class CreateBingsooCommand (
     val userId: Long,
     val taste: Taste
 )
